backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
spacy==3.7.2
pymupdf==1.23.8
peft==0.7.1
bitsandbytes==0.41.3
accelerate==0.24.1
huggingface-hub==0.19.4 